chore: move directory of inventory/job-task
