+++ apps/web/src/services/asset-inventory/pages/admin/AdminCollectHistoryJobPage.vue
@@ -10,8 +10,8 @@ import {
     PHorizontalLayout, PHeading,
 } from '@cloudforet/mirinae';
 
+import type { JobTaskModel } from '@/api-clients/inventory/job-task/schema/model';
 import type { JobModel } from '@/api-clients/inventory/job/schema/model';
-import type { JobTaskModel } from '@/schema/inventory/job-task/model';
 
 import ErrorHandler from '@/common/composables/error/errorHandler';
